+++ test/torus/query_inspector_test.exs
@@ -4,7 +4,7 @@ defmodule Torus.QueryInspectorTest do
   alias Torus.Test.Repo
 
   test "substituted_sql/2 - default" do
-    assert "SELECT p0.\"id\", p0.\"title\", p0.\"body\", p0.\"author_id\" FROM \"posts\" AS p0 WHERE (p0.\"id\" = 1)" =
+    assert "SELECT p0.\"id\", p0.\"title\", p0.\"body\", p0.\"embedding\", p0.\"author_id\" FROM \"posts\" AS p0 WHERE (p0.\"id\" = 1)" =
              Post |> where(id: 1) |> Torus.QueryInspector.substituted_sql(Repo)
   end
 
@@ -12,7 +12,7 @@ defmodule Torus.QueryInspectorTest do
     int_array = [1, 2]
     binary_array = ["hello", "world"]
 
-    assert "SELECT p0.\"id\", p0.\"title\", p0.\"body\", p0.\"author_id\" FROM \"posts\" AS p0 WHERE (p0.\"id\" = ANY(ARRAY[1,2])) AND (p0.\"title\" = ANY(ARRAY['hello','world'])) AND ((p0.\"title\" ILIKE 'test%') OR 'false')" =
+    assert "SELECT p0.\"id\", p0.\"title\", p0.\"body\", p0.\"embedding\", p0.\"author_id\" FROM \"posts\" AS p0 WHERE (p0.\"id\" = ANY(ARRAY[1,2])) AND (p0.\"title\" = ANY(ARRAY['hello','world'])) AND ((p0.\"title\" ILIKE 'test%') OR 'false')" =
              Post
              |> where([p], p.id in ^int_array)
              |> where([p], p.title in ^binary_array)